[FRONT] restructure the catalogs view and fix a bug in create condition controller
